+++ example/ToDo List/Server/RouteDefenitions.dart
@@ -11,20 +11,20 @@ var searchQuery = new QVariable("q");
 // ----------Route Defenition - Page----------
 
 // Gobal Pathes
-var jsPath = new RouteDef("js/*", "JS File access");
-var dartPath = new RouteDef("dart/*", "Dart File access");
-var cssPath = new RouteDef("css/*", "CSS File access");
+var jsPath = new Url("js/*", "JS File access");
+var dartPath = new Url("dart/*", "Dart File access");
+var cssPath = new Url("css/*", "CSS File access");
 
 // Pages
-var jshome = new RouteDef("", "Home JavaScript");
-var darthome = new RouteDef("darthome", "Home Dart");
-var about = new RouteDef.fromObjects([new Static("about")], name: "Aboute Route");
-var todoLists = new RouteDef.fromObjects([toDo], name: "ToDo Lists Route");
-var todoList = new RouteDef.fromObjects([toDo, listID], name: "ToDo List Route");
-var todoItem = new RouteDef.fromObjects([toDo, listID, itemID], name: "ToDo Item Route");
-var search = new RouteDef.fromObjects([new Static("search")], queryParts: [searchQuery], name: "ToDo Item Route");
+var jshome = new Url("", "Home JavaScript");
+var darthome = new Url("darthome", "Home Dart");
+var about = new Url.fromObjects([new Static("about")], name: "Aboute Route");
+var todoLists = new Url.fromObjects([toDo], name: "ToDo Lists Route");
+var todoList = new Url.fromObjects([toDo, listID], name: "ToDo List Route");
+var todoItem = new Url.fromObjects([toDo, listID, itemID], name: "ToDo Item Route");
+var search = new Url.fromObjects([new Static("search")], queryParts: [searchQuery], name: "ToDo Item Route");
 
 // ----------Route Defenition - API----------
-var api_todoLists = new RouteDef.fromObjects([version, toDo], name: "API: ToDo Lists Route");
-var api_todoList = new RouteDef.fromObjects([version, toDo, listID], name: "API: ToDo List Route");
-var api_todoItem = new RouteDef.fromObjects([version, toDo, listID, itemID], name: "API: ToDo Item Route");
+var api_todoLists = new Url.fromObjects([version, toDo], name: "API: ToDo Lists Route");
+var api_todoList = new Url.fromObjects([version, toDo, listID], name: "API: ToDo List Route");
+var api_todoItem = new Url.fromObjects([version, toDo, listID, itemID], name: "API: ToDo Item Route");