+++ lib/BullsEyeCommon/src/Extensions/version_path_part.dart
@@ -9,12 +9,12 @@ class Version extends Variable {
     RegisterParseFunc();
   }
   
-  void CheckUsage(RouteDef currentRoute) 
+  void CheckUsage(Url currentRoute) 
   {
     if(currentRoute.routeParts.where((part) => part is Version).length > 1) throw new MultipleVersionException(currentRoute);
   }
   
-  static Variable _TryParse(String rawStr, String cleanStr, bool isOptional, RouteDefConfig currentConfig, RouteDef currentRoute) {
+  static Variable _TryParse(String rawStr, String cleanStr, bool isOptional, UrlDefConfig currentConfig, Url currentRoute) {
     
     if(cleanStr == routePartVariableVersion && !isOptional)
     {
@@ -28,7 +28,7 @@ class Version extends Variable {
   
   static void RegisterParseFunc()
   {
-    var config = new RouteDefConfig.Current();
+    var config = new UrlDefConfig.Current();
     if(!config.specialVariableParsers.contains(_TryParse))
     {
       config.RegisterSpecialVariableParser(_TryParse);
@@ -38,7 +38,7 @@ class Version extends Variable {
 
 class MultipleVersionException implements Exception {
 
-  RouteDef _routeDefinition;
+  Url _routeDefinition;
   
   MultipleVersionException(this._routeDefinition);