+++ lib/BullsEyeCommon/src/RouteDefenition/DefExceptions.dart
@@ -1,2 +1,0 @@
-part of softhai.bulls_eye.Common;
-