+++ lib/BullsEyeServer/bulls_eye_server.dart
@@ -3,9 +3,18 @@ library softhai.bulls_eye.Server;
 import 'dart:io';
 import '../BullsEyeCommon/bulls_eye_common.dart' as common;
 
-part 'src/Routing/RoutingExceptions.dart';
+// Globals
+part 'src/req_res_context.dart';
 part 'src/Server.dart';
 part 'src/Exceptions.dart';
+
+// Middleware
+part 'src/Middleware/Middleware.dart';
+part 'src/Middleware/middleware_channel_part.dart';
+part 'src/Middleware/Middleware_error.dart';
+
+// Routing
 part 'src/Routing/Route.dart';
 part 'src/Routing/LogicRoute.dart';
 part 'src/Routing/FileRoute.dart';
+part 'src/Routing/RoutingExceptions.dart';