+++ lib/BullsEyeServer/src/Middleware/middleware_error.dart
@@ -1,0 +1,10 @@
+part of softhai.bulls_eye.Server;
+
+class MiddlewareError {
+  
+  String MiddlewareName;
+  Object catchedError;
+  
+  MiddlewareError(this.MiddlewareName, this.catchedError);
+  
+}